+++ src/frequenz/channels/experimental/_with_previous.py
@@ -1,7 +1,7 @@
 # License: MIT
 # Copyright © 2024 Frequenz Energy-as-a-Service GmbH
 
-"""Predicates to be used in conjuntion with `Receiver.filter()`."""
+"""Composable predicate to cache and compare with the previous message."""
 
 
 from typing import Callable, Final, Generic, TypeGuard
@@ -20,48 +20,43 @@ class _Sentinel:
 _SENTINEL: Final[_Sentinel] = _Sentinel()
 
 
-class OnlyIfPrevious(Generic[ChannelMessageT]):
-    """A predicate to check if a message has a particular relationship with the previous one.
+class WithPrevious(Generic[ChannelMessageT]):
+    """A composable predicate to build predicates that can use also the previous message.
 
-    This predicate can be used to filter out messages based on a custom condition on the
+    This predicate can be used to filter messages based on a custom condition on the
     previous and current messages. This can be useful in cases where you want to
     process messages only if they satisfy a particular condition with respect to the
     previous message.
 
-    Tip:
-        If you want to use `==` as predicate, you can use the
-        [`ChangedOnly`][frequenz.channels.experimental.ChangedOnly] predicate.
-
-    Example: Receiving only messages that are not the same instance as the previous one.
+    Example: Receiving only messages that are different from the previous one.
         ```python
         from frequenz.channels import Broadcast
-        from frequenz.channels.experimental import OnlyIfPrevious
+        from frequenz.channels.experimental import WithPrevious
 
-        channel = Broadcast[int | bool](name="example")
-        receiver = channel.new_receiver().filter(OnlyIfPrevious(lambda old, new: old is not new))
+        channel = Broadcast[int](name="example")
+        receiver = channel.new_receiver().filter(WithPrevious(lambda old, new: old != new))
         sender = channel.new_sender()
 
         # This message will be received as it is the first message.
         await sender.send(1)
         assert await receiver.receive() == 1
 
-        # This message will be skipped as it is the same instance as the previous one.
+        # This message will be skipped as it equals to the previous one.
         await sender.send(1)
 
-        # This message will be received as it is a different instance from the previous
-        # one.
-        await sender.send(True)
-        assert await receiver.receive() is True
+        # This message will be received as it is a different from the previous one.
+        await sender.send(0)
+        assert await receiver.receive() == 0
         ```
 
     Example: Receiving only messages if they are bigger than the previous one.
         ```python
         from frequenz.channels import Broadcast
-        from frequenz.channels.experimental import OnlyIfPrevious
+        from frequenz.channels.experimental import WithPrevious
 
         channel = Broadcast[int](name="example")
         receiver = channel.new_receiver().filter(
-            OnlyIfPrevious(lambda old, new: new > old, first_is_true=False)
+            WithPrevious(lambda old, new: new > old, first_is_true=False)
         )
         sender = channel.new_sender()
 
@@ -90,6 +85,7 @@ class OnlyIfPrevious(Generic[ChannelMessageT]):
     def __init__(
         self,
         predicate: Callable[[ChannelMessageT, ChannelMessageT], bool],
+        /,
         *,
         first_is_true: bool = True,
     ) -> None:
@@ -127,58 +123,3 @@ class OnlyIfPrevious(Generic[ChannelMessageT]):
     def __repr__(self) -> str:
         """Return a string representation of this instance."""
         return f"<{type(self).__name__}: {self._predicate!r} first_is_true={self._first_is_true!r}>"
-
-
-class ChangedOnly(OnlyIfPrevious[object]):
-    """A predicate to check if a message is different from the previous one.
-
-    This predicate can be used to filter out messages that are the same as the previous
-    one. This can be useful in cases where you want to avoid processing duplicate
-    messages.
-
-    Warning:
-        This predicate uses the `!=` operator to compare messages, which includes all
-        the weirdnesses of Python's equality comparison (e.g., `1 == 1.0`, `True == 1`,
-        `True == 1.0`, `False == 0` are all `True`).
-
-        If you need to use a different comparison, you can create a custom predicate
-        using [`OnlyIfPrevious`][frequenz.channels.experimental.OnlyIfPrevious].
-
-    Example:
-        ```python
-        from frequenz.channels import Broadcast
-        from frequenz.channels.experimental import ChangedOnly
-
-        channel = Broadcast[int](name="skip_duplicates_test")
-        receiver = channel.new_receiver().filter(ChangedOnly())
-        sender = channel.new_sender()
-
-        # This message will be received as it is the first message.
-        await sender.send(1)
-        assert await receiver.receive() == 1
-
-        # This message will be skipped as it is the same as the previous one.
-        await sender.send(1)
-
-        # This message will be received as it is different from the previous one.
-        await sender.send(2)
-        assert await receiver.receive() == 2
-        ```
-    """
-
-    def __init__(self, *, first_is_true: bool = True) -> None:
-        """Initialize this instance.
-
-        Args:
-            first_is_true: Whether the first message should be considered as different
-                from the previous one. Defaults to `True`.
-        """
-        super().__init__(lambda old, new: old != new, first_is_true=first_is_true)
-
-    def __str__(self) -> str:
-        """Return a string representation of this instance."""
-        return f"{type(self).__name__}"
-
-    def __repr__(self) -> str:
-        """Return a string representation of this instance."""
-        return f"{type(self).__name__}(first_is_true={self._first_is_true!r})"