+++ RELEASE_NOTES.md
@@ -4,5 +4,4 @@
 
 ### Experimental
 
-- A new predicate, `OnlyIfPrevious`, to `filter()` messages based on the previous message.
-- A new special case of `OnlyIfPrevious`, `ChangedOnly`, to skip messages if they are equal to the previous message.
+- A new composable predicate, `WithPrevious`, to filter messages based on the previous message.